refactor: remove obsolete authentication and user role services; implement new user role fetching logic
