--- src/app/api/auth/login/route.ts
@@ -1,49 +1,0 @@
-import { prisma } from "@/lib/prisma";
-import { NextRequest, NextResponse } from "next/server";
-import bcrypt from "bcrypt";
-
-export async function POST(req: NextRequest): Promise<NextResponse> {
-  try {
-    const { email, password } = await req.json();
-
-    const user = await prisma.user.findUnique({
-      where: { email },
-    });
-
-    if (!user) {
-      return NextResponse.json(
-        { error: "Email not found. Please register first" },
-        { status: 404 }
-      );
-    }
-
-    if (!user.password) {
-      return NextResponse.json(
-        { error: "Invalid email or password" },
-        { status: 401 }
-      );
-    }
-
-    const isPasswordValid = await bcrypt.compare(password, user.password);
-
-    if (!isPasswordValid) {
-      return NextResponse.json(
-        { error: "Invalid email or password" },
-        { status: 401 }
-      );
-    }
-
-    return NextResponse.json(
-      { message: "Login successful", user: { id: user.id, firstName: user.firstName, lastName: user.lastName, email: user.email } },
-      { status: 200 }
-    );
-  } catch (err) {
-    console.error("Login error:", err);
-    return NextResponse.json(
-      { error: "An error occurred during login" },
-      { status: 500 }
-    );
-  } finally {
-    await prisma.$disconnect();
-  }
-}
