--- src/domains/user-management/services/getUserDepartment.ts
@@ -1,10 +1,0 @@
-import { prisma } from "@/lib/prisma";
-
-export default async function getUserDepartment(userId: string): Promise<string> {
-  const user = await prisma.user.findUnique({
-    where: { id: userId },
-    select: { department: true },
-  });
-
-  return user?.department ?? "";
-}
